+++ src/components/BotDetails/UserInfoRow.js
@@ -39,7 +39,11 @@ class UserInfoRow extends React.Component {
     const profile = owner;
     return (
       <View style={styles.container}>
-        <RText color={colors.DARK_PURPLE} numberOfLines={2} size={18}>{`${bot.title}`}</RText>
+        <View style={{flexDirection: 'row', alignItems: 'center'}}>
+          <RText color={colors.DARK_PURPLE} numberOfLines={2} size={18}>{`${bot.title}`}</RText>
+          {!bot.isPublic && <Image source={require('../../../images/iconPrivate.png')} style={{marginHorizontal: 7 * k}} resizeMode='contain' />}
+        </View>
+
         <View style={styles.userInfoRow}>
           <ProfileAvatar profile={profile} size={40 * k} />
           <View style={{marginLeft: 10 * k, flex: 1}}>
@@ -47,7 +51,7 @@ class UserInfoRow extends React.Component {
               <RText weight='Medium' size={15} color={colors.DARK_PURPLE}>{`@${profile.handle}`}</RText>
             </TouchableOpacity>
           </View>
-          <Image style={{width: 14 * k, height: 13 * k}} source={require('../../../images/heart.png')} />
+          <SavesCount botId={bot.id} isOwn={owner && owner.isOwn} />
           <RText color={colors.WARM_GREY_2} style={{marginLeft: 4 * k, marginRight: 4 * k}}>
             {bot.followersSize}
           </RText>
@@ -73,6 +77,17 @@ class UserInfoRow extends React.Component {
     );
   }
 }
+
+const SavesCount = ({botId, isOwn}) => {
+  const inner = <Image style={{width: 14 * k, height: 13 * k}} source={require('../../../images/heart.png')} />;
+  return isOwn ? (
+    <TouchableOpacity onPress={() => Actions.subscribers({item: botId})} hitSlop={{top: 5, right: 5, bottom: 5, left: 5}}>
+      {inner}
+    </TouchableOpacity>
+  ) : (
+    inner
+  );
+};
 
 export default UserInfoRow;