+++ src/components/people-lists/index.js
@@ -3,5 +3,6 @@ import BlockedList from './BlockedList';
 import FriendListScene from './FriendListScene';
 import FollowersList from './FollowersList';
 import FollowingList from './FollowingList';
+import BotSubscriberList from './BotSubscriberList';
 
-export {BlockedList, FriendListScene, FollowersList, FollowingList};
+export {BlockedList, FriendListScene, FollowersList, FollowingList, BotSubscriberList};